move select functions to lib/selections.ts to avoid circular deps
